[Bronze V] Title: 이상한 기호, Time: 48 ms, Memory: 5508 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cs
@@ -14,7 +14,7 @@ namespace code
             StreamReader sr = new StreamReader(Console.OpenStandardInput());
             StringBuilder sb = new StringBuilder();
 
-            int[] inputs = Array.ConvertAll(sr.ReadLine().Split(' '), int.Parse);
+            long[] inputs = Array.ConvertAll(sr.ReadLine().Split(' '), long.Parse);
 
             sb.Append((inputs[0] + inputs[1]) * (inputs[0] - inputs[1]));
